refactor(test-framework): Move the CommandLineBuilder and VersionParser utilities to the common package
