+++ src/StaticAnalysis/Mago/Adapter/MagoAdapter.php
@@ -41,8 +41,8 @@ use Infection\Process\Factory\LazyMutantProcessFactory;
 use Infection\Process\ShellCommandLineExecutor;
 use Infection\StaticAnalysis\Mago\Process\MagoMutantProcessFactory;
 use Infection\StaticAnalysis\StaticAnalysisToolAdapter;
-use Infection\TestFramework\CommandLineBuilder;
-use Infection\TestFramework\VersionParser;
+use Infection\TestFramework\Common\CommandLineBuilder;
+use Infection\TestFramework\Common\VersionParser;
 use RuntimeException;
 use Safe\Exceptions\PcreException;
 use function sprintf;
