+++ src/StaticAnalysis/Mago/Process/MagoMutantProcessFactory.php
@@ -40,7 +40,7 @@ use Infection\Mutant\Mutant;
 use Infection\Mutant\MutantExecutionResultFactory;
 use Infection\Process\Factory\LazyMutantProcessFactory;
 use Infection\Process\MutantProcess;
-use Infection\TestFramework\CommandLineBuilder;
+use Infection\TestFramework\Common\CommandLineBuilder;
 use Symfony\Component\Process\Process;
 
 /**
