+++ src/StaticAnalysis/PHPStan/Adapter/PHPStanAdapter.php
@@ -42,8 +42,8 @@ use Infection\Process\Factory\LazyMutantProcessFactory;
 use Infection\Process\ShellCommandLineExecutor;
 use Infection\StaticAnalysis\PHPStan\Process\PHPStanMutantProcessFactory;
 use Infection\StaticAnalysis\StaticAnalysisToolAdapter;
-use Infection\TestFramework\CommandLineBuilder;
-use Infection\TestFramework\VersionParser;
+use Infection\TestFramework\Common\CommandLineBuilder;
+use Infection\TestFramework\Common\VersionParser;
 use RuntimeException;
 use function sprintf;
 use function str_starts_with;
