+++ src/TestFramework/AbstractTestFrameworkAdapter.php
@@ -38,6 +38,8 @@ namespace Infection\TestFramework;
 use Infection\AbstractTestFramework\Coverage\TestLocation;
 use Infection\AbstractTestFramework\TestFrameworkAdapter;
 use Infection\Process\ShellCommandLineExecutor;
+use Infection\TestFramework\Common\CommandLineBuilder;
+use Infection\TestFramework\Common\VersionParser;
 use Infection\TestFramework\Config\InitialConfigBuilder;
 use Infection\TestFramework\Config\MutationConfigBuilder;
 use function sprintf;
