+++ src/TestFramework/Common/VersionParser.php
@@ -33,7 +33,7 @@
 
 declare(strict_types=1);
 
-namespace Infection\TestFramework;
+namespace Infection\TestFramework\Common;
 
 use function Safe\preg_match;
 use function sprintf;
