+++ src/TestFramework/PhpUnit/Adapter/PhpUnitAdapter.php
@@ -43,11 +43,11 @@ use Infection\Config\ValueProvider\PCOVDirectoryProvider;
 use Infection\Process\ShellCommandLineExecutor;
 use Infection\TestFramework\AbstractTestFrameworkAdapter;
 use Infection\TestFramework\CommandLineArgumentsAndOptionsBuilder;
-use Infection\TestFramework\CommandLineBuilder;
+use Infection\TestFramework\Common\CommandLineBuilder;
+use Infection\TestFramework\Common\VersionParser;
 use Infection\TestFramework\Config\InitialConfigBuilder;
 use Infection\TestFramework\Config\MutationConfigBuilder;
 use Infection\TestFramework\ProvidesInitialRunOnlyOptions;
-use Infection\TestFramework\VersionParser;
 use Override;
 use function Safe\preg_match;
 use function sprintf;
